# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Global Home Stub Api's",
    "description": "Backend API's for GlobalHome.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/properties/units/{unitId}": {
      "get": {
        "tags": [
          "units"
        ],
        "summary": "Get individual units detail associated with this Property.",
        "description": "Get individual units detail associated with this Property",
        "operationId": "getUnitByid",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "description": "The ID of the unit.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccess"
                },
                "example": {
                  "statusCode": 200,
                  "status": "success",
                  "message": "Unit info  successfully.",
                  "data": {
                    "id": "234324242423",
                    "name": "test name",
                    "unit_type": "building",
                    "property_id": "2133214",
                    "building_id": "243243",
                    "physical_location": "Delhi",
                    "updatedAt": "2020-09-08T06:33:32.458Z",
                    "createdAt": "2020-09-08T06:33:32.458Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Access not authorized or forbidden",
            "content": {}
          },
          "404": {
            "description": "No information found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "units"
        ],
        "summary": "Update individual units detail associated with this Property.",
        "description": "Update individual units detail associated with this Property",
        "operationId": "updateUnit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "description": "The ID of the unit.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated unit request body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccess"
                },
                "example": {
                  "statusCode": 200,
                  "status": "success",
                  "message": "Unit updated successfully.",
                  "data": {
                    "id": "234324242423",
                    "name": "test name",
                    "unit_type": "building",
                    "property_id": "2133214",
                    "building_id": "243243",
                    "physical_location": "Delhi",
                    "updatedAt": "2020-09-08T06:33:32.458Z",
                    "createdAt": "2020-09-08T06:33:32.458Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Access not authorized or forbidden",
            "content": {}
          },
          "422": {
            "description": "Validation faliled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "units"
        ],
        "summary": "Delete unit detail",
        "description": "Delete unit detail",
        "operationId": "deleteUnit",
        "parameters": [
          {
            "name": "unitId",
            "in": "path",
            "description": "The ID of the unit.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseSuccess"
                },
                "example": {
                  "statusCode": 200,
                  "status": "success",
                  "message": "Unit detail deleted successfully.",
                  "data": null
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Access not authorized or forbidden",
            "content": {}
          },
          "500": {
            "description": "Error on delete operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "Status code 422"
          },
          "status": {
            "type": "string",
            "description": "Error or success"
          },
          "message": {
            "type": "array",
            "description": "array of falied validation message"
          },
          "data": {
            "type": "object",
            "description": "Data object comes null in case of error"
          }
        },
        "example": {
          "statusCode": 422,
          "status": "error",
          "message": [
            "The name field is mandatory."
          ],
          "data": null
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "Status code 422"
          },
          "status": {
            "type": "string",
            "description": "Error or success"
          },
          "message": {
            "type": "array",
            "description": "array of falied validation message"
          },
          "data": {
            "type": "object",
            "description": "Data object comes null in case of error"
          }
        },
        "example": {
          "statusCode": 500,
          "status": "error",
          "message": "Something went wrong",
          "data": null
        }
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "Status code 404"
          },
          "status": {
            "type": "string",
            "description": "Error or success"
          },
          "message": {
            "type": "string",
            "description": "No information found"
          },
          "data": {
            "type": "object",
            "description": "Data object comes null in case of error"
          }
        },
        "example": {
          "statusCode": 404,
          "status": "error",
          "message": "Data not found ",
          "data": null
        }
      },
      "ResponseSuccess": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "Status code 200"
          },
          "status": {
            "type": "string",
            "description": "success"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Data object comes null in case of error"
          }
        }
      },
...
      "UpdateUnitRequest": {
        "required": [
          "id",
          "name",
          "unit_type",
          "physical_location"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "required": true
          },
          "name": {
            "type": "string",
            "required": true
          },
          "unit_type": {
            "type": "string",
            "required": true
          },
          "physical_location": {
            "type": "string",
            "required": true
          }
        },
        "example": {
          "id": "ASDADS-SDDSA-ADSADS-ASDDSA",
          "name": "Apartment 532A",
          "unit_type": "Residential",
          "physical_location": "Apartment 532A Apartment 532A "
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}